Fix for "Invalid Data Format" Error when pulling Training Data for 'keyword spotting' Project
The @totalLengthMs was a 'Number', so not an 'Integer'.

+++ Shared/State/DataSample.swift
@@ -13,10 +13,10 @@ struct DataSample: Identifiable, Codable {
     let filename: String
     let category: Category
     let label: String
-    let totalLengthMs: Int
+    let totalLengthMs: Double
     
     func totalLengthInSeconds() -> String {
-        let double = Measurement(value: Double(totalLengthMs), unit: UnitDuration.milliseconds)
+        let double = Measurement(value: totalLengthMs, unit: UnitDuration.milliseconds)
             .converted(to: .seconds).value
         return String(format: "%.2fs", double)
     }
